refactor: modularize core commands
