+++ core/__init__.py
@@ -1,3 +1,2 @@
 from .bot import Bot
 from .player import Player
-from .commands import Command
