+++ core/commands/__init__.py
@@ -1,0 +1,20 @@
+﻿from .base import _CommandBase, check_alive
+from .quest_commands import QuestCommands
+from .item_commands import ItemCommands
+from .map_commands import MapCommands
+from .combat_commands import CombatCommands
+from .player_commands import PlayerCommands
+from .utility_commands import UtilityCommands
+from ..command import Command
+
+__all__ = [
+    "_CommandBase",
+    "check_alive",
+    "QuestCommands",
+    "ItemCommands",
+    "MapCommands",
+    "CombatCommands",
+    "PlayerCommands",
+    "UtilityCommands",
+    "Command",
+]
